backend/requirements.txt
fastapi==0.68.1
python-multipart==0.0.5
pillow==8.3.2
opencv-python==4.5.3.56
numpy==1.21.2
uvicorn==0.15.0
python-jose==3.3.0 